Break apart the create wallet styles
