+++ src/components/scenes/CreateWalletSelectFiatScene.js
@@ -1,7 +1,7 @@
 // @flow
 
 import React, { Component } from 'react'
-import { Alert, FlatList, TouchableHighlight, View } from 'react-native'
+import { Alert, FlatList, StyleSheet, TouchableHighlight, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import { connect } from 'react-redux'
 
@@ -9,7 +9,7 @@ import * as Constants from '../../constants/indexConstants.js'
 import s from '../../locales/strings.js'
 import { getDefaultFiat } from '../../modules/Settings/selectors.js'
 import Text from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import styles, { styles as stylesRaw } from '../../styles/scenes/CreateWalletStyle.js'
+import { THEME } from '../../theme/variables/airbitz.js'
 import { type Dispatch, type State as ReduxState } from '../../types/reduxTypes.js'
 import type { CreateWalletType, FlatListItem, GuiFiatType } from '../../types/types.js'
 import { scale } from '../../util/scaling.js'
@@ -142,7 +142,7 @@ class CreateWalletSelectFiatComponent extends Component<Props, State> {
   renderFiatTypeResult = (data: FlatListItem<GuiFiatType>) => {
     return (
       <View style={[styles.singleCryptoTypeWrap, data.item.value === this.state.selectedFiat && styles.selectedItem]}>
-        <TouchableHighlight style={styles.singleCryptoType} onPress={() => this.handleSelectFiatType(data.item)} underlayColor={stylesRaw.underlayColor.color}>
+        <TouchableHighlight style={styles.singleCryptoType} onPress={() => this.handleSelectFiatType(data.item)} underlayColor={THEME.COLORS.GRAY_4}>
           <View style={styles.cryptoTypeInfoWrap}>
             <View style={styles.cryptoTypeLeft}>
               <View style={styles.cryptoTypeLeftTextWrap}>
@@ -159,6 +159,61 @@ class CreateWalletSelectFiatComponent extends Component<Props, State> {
     return item.value
   }
 }
+
+const rawStyles = {
+  content: {
+    backgroundColor: THEME.COLORS.WHITE,
+    flex: 1,
+    paddingHorizontal: scale(20)
+  },
+  picker: {
+    fontFamily: THEME.FONTS.DEFAULT,
+    height: scale(50),
+    padding: scale(5)
+  },
+  resultList: {
+    backgroundColor: THEME.COLORS.WHITE,
+    borderTopColor: THEME.COLORS.GRAY_3,
+    borderTopWidth: 1,
+    flex: 1
+  },
+  selectedItem: {
+    backgroundColor: THEME.COLORS.GRAY_4,
+    borderLeftWidth: scale(1),
+    borderLeftColor: THEME.COLORS.GRAY_3,
+    borderRightWidth: scale(1),
+    borderRightColor: THEME.COLORS.GRAY_3
+  },
+  singleCryptoType: {
+    height: scale(60),
+    borderBottomWidth: scale(1),
+    borderBottomColor: THEME.COLORS.GRAY_3,
+    paddingVertical: scale(10),
+    paddingHorizontal: scale(15)
+  },
+  singleCryptoTypeWrap: {
+    flexDirection: 'column',
+    flex: 1
+  },
+  cryptoTypeInfoWrap: {
+    flexDirection: 'row',
+    height: scale(40),
+    flex: 1,
+    justifyContent: 'space-between'
+  },
+  cryptoTypeLeft: {
+    flexDirection: 'row'
+  },
+  cryptoTypeLeftTextWrap: {
+    justifyContent: 'center'
+  },
+  cryptoTypeName: {
+    fontSize: scale(16),
+    color: THEME.COLORS.GRAY_1,
+    textAlignVertical: 'center'
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
 
 export const CreateWalletSelectFiatScene = connect(
   (state: ReduxState): StateProps => ({
